Move db-related functions used on /graafi to database.js
No functional change
